Arreglo los patch de los tests
