+++ django_datajsonar/models/__init__.py
@@ -1,3 +1,5 @@
+#! coding: utf-8
+from .utils import filepath, get_distribution_storage
 from .data_json import Catalog, Dataset, Distribution, Field
 from .metadata import Metadata
 from .tasks import ReadDataJsonTask
