+++ django_datajsonar/models/metadata.py
@@ -17,4 +17,3 @@ class Metadata(models.Model):
     content_type = models.ForeignKey(ContentType, on_delete=models.CASCADE)
     object_id = models.PositiveIntegerField()
     content_object = GenericForeignKey()
-
